requirements.txt
certifi==2019.9.11
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.4
djangorestframework==3.10.3
gunicorn==19.9.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.3
Markdown==3.1.1
MarkupSafe==1.1.1
Pygments==2.4.2
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.6
whitenoise==4.1.4
